+++ tests/error/errors_on_usage/var_in_modifier3.py
@@ -1,0 +1,24 @@
+
+from guppylang import guppy, qubit
+from guppylang.std.builtins import owned, power, control
+from guppylang.std.quantum import measure
+
+
+@guppy
+def main() -> int:
+    q = qubit()
+    if True:
+        x = 3
+    with power(3): 
+        b = 4
+        with control(q):
+            if b:
+                a = 4
+    
+    measure(q)
+
+    return a
+
+
+
+main.check()