+++ tests/error/modifier_errors/flag_dagger_qubit1.py
@@ -4,7 +4,7 @@ from guppylang.std.builtins import owned
 @guppy.declare(unitary=True)
 def uni_discard(q: qubit @owned) -> None: ...
 
-@guppy(dagger=True)
+@guppy(daggerable=True)
 def test() -> None:
     p = qubit()
     uni_discard(p)